requirements.txt
annotated-types==0.7.0
boto3==1.34.149
botocore==1.34.149
click==8.1.7
coverage==7.6.0
gevent==24.2.1
greenlet==3.0.3
iniconfig==2.0.0
Jinja2==3.1.3
jmespath==1.0.1
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==23.2
pillow==10.3.0
pluggy==1.5.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pytest==8.1.1
pytest-cov==4.1.0
pytest-mock==3.3.1
python-dateutil==2.9.0.post0
python-pptx==0.6.23
PyYAML==6.0.1
rich==13.7.0
s3transfer==0.10.2
six==1.16.0
typing_extensions==4.9.0
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.0.1
XlsxWriter==3.2.0
zope.event==5.0
zope.interface==7.0.3
